Add functionality to handle script-built events.
When scripts built entities, it is up to them to actually fire a script_raised_built event, but when they do, we want to grab it.  If the content is similar (read, has created_entity + player_index) to a regular on_built_entity event, then we'll pass that along as if it were a regular player building the entity.

Also implemented fix for building the GUI button to disable autofill when added in as a mod after the start of the game.