Fixes #6.  Force to fuel/ammo inventories, work around race condition.
This fix forces the insert for fuels & ammos to be put into their respective inventories, which means that this mod won't cause an overflow into a vehicle's "trunk".  Since other mods might be inserting fuel/ammo into the same entity we are, then we also handle the case where we're not first to this entity, and only insert a reduced ammount (limited to the inventoy insert).

Additionally, to make this work with other mods that are not going to this extra length on inserting fuel/ammo, we'll make sure to push our 'on-script-built' entity insert to the next tick, so that we can work around the issue of multiple mods inserting (in various orders) fuel & ammo into the created entity.  See the Vehicle Wagon mod for an example of one that does this insert as well.