Add loop detection for ammo damage calculation
Some mods (*cough*Space Exploration*cough*) had some damage effects which would create a damage entity (1) that would create a damage entity (2) that created damage entity (1) and back & forth infinitely.  So, we'll check that we're not adding the effect entities in a loop, and throw out a warning because the mod developers doing this ought to fix their crap.